% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BhattDist}
\alias{BhattDist}
\title{Bhattacharyya distance}
\usage{
BhattDist(mu1, S1, mu2, S2)
}
\arguments{
\item{mu1}{mean vector of P.}

\item{S1}{covariance matrix of P.}

\item{mu2}{mean vector of Q.}

\item{S2}{covariance matrix of Q.}
}
\value{
Bhattacharyya distance between P and Q.
}
\description{
Bhattacharyya distance between P and Q for multivariate normal distributions
}
\keyword{internal}
