% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compile.h}
\alias{compile.h}
\title{Generate function for hazard}
\usage{
compile.h(rct.lst, envir)
}
\arguments{
\item{rct.lst}{list of biochemical reactions.
A differentiation move from cell type "A" to cell type "B" must be coded as "A->B"
Duplication of cell "A" must be coded as "A->1"
Death of cell "A" must be coded as "A->0"}
}
\value{
The function doesn't return anything.
}
\description{
This function dynamically builds the function get.h() that will be used to get the hazard vector.
}
\keyword{internal}
