% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dBu}
\alias{dBu}
\title{Gradient of \eqn{E[u \vert \theta]}{E[u|θ]}}
\usage{
dBu(K, p)
}
\arguments{
\item{K}{number of clones being analyzed.}

\item{p}{number of free parameters.}
}
\value{
the p-dimensional gradient of u.
}
\description{
Gradient of the expected values of the random effects u w.r.t. the free parameters
}
\keyword{internal}
