% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dD2}
\alias{dD2}
\title{Gradient of \eqn{V[u \vert \theta]}{V[u|θ]}}
\usage{
dD2(K, p)
}
\arguments{
\item{K}{number of clones being analyzed.}

\item{p}{number of free parameters.}
}
\value{
the gradient of the covariance matrix of u.
}
\description{
Gradient of the covariance matrix of the random effects u w.r.t. the free parameters
}
\keyword{internal}
