% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dmvNorm}
\alias{dmvNorm}
\title{Multivariate normal density function}
\usage{
dmvNorm(x, Mu, Sigma, SigmaInv)
}
\arguments{
\item{x}{n-dimensional vector.}

\item{Mu}{n-dimensional mean vector.}

\item{Sigma}{n-dimensional covariance matrix.}

\item{SigmaInv}{n-dimensional precision matrix.}
}
\value{
the probability density of a n-variate normal distribution with mean vector Mu and covariance Sigma evaluated at x.
}
\description{
Proability density function of a n-variate normal distribution
}
\keyword{internal}
