% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.dx}
\alias{get.dx}
\title{Time-adjacent increments of the cell counts}
\usage{
get.dx(y)
}
\arguments{
\item{y}{clone-specific \eqn{t \times p}{t by p} cell count matrix, where t is the number of time-points and p the number of cell types.}
}
\value{
A \eqn{(t-1) \times p}{(t - 1) by p} dimensional vector containing the time-adjacent increments.
}
\description{
This function generates time-adjacent increments from a cell counts matrix y.
}
\keyword{internal}
