% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.nl}
\alias{get.nl}
\title{Negative log-likelihood of the base (null) model}
\usage{
get.nl(theta, M, y, V, VCNs, nObs, dW)
}
\arguments{
\item{theta}{p-dimensional vector parameter.}

\item{M}{A \eqn{n \times K}{n by K} dimensional (design) matrix.}

\item{y}{n-dimensional vector of the time-adjacent cellular increments}

\item{V}{A \eqn{p \times K}{p by K} dimensional net-effect matrix.}

\item{dW}{p-dimensional list of the partial derivatives of W w.r.t. theta.}
}
\value{
the value of the negative log-likelihood.
}
\description{
Compute the negative log-likelihood of the null model,
namely the linear model without rondom effects at all.
}
\keyword{internal}
