% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.V}
\alias{get.V}
\title{Net-effect matrix}
\usage{
get.V(ct.lst, rct.lst)
}
\arguments{
\item{ct.lst}{list of cell types}

\item{rct.lst}{list of biochemical reactions.
A differentiation move from cell type "A" to cell type "B" must be coded as "A->B"
Duplication of cell "A" must be coded as "A->1"
Death of cell "A" must be coded as "A->0"}
}
\value{
The net-effect matrix V
}
\description{
This function builds the net-effect matrix V.
}
\keyword{internal}
