% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.W}
\alias{get.W}
\title{Stochastic covariance matrix W}
\usage{
get.W(M, V, VCNs, theta, nObs)
}
\arguments{
\item{M}{A \eqn{n \times K}{n by K} dimensional (design) matrix.}

\item{V}{A \eqn{p \times K}{p by K} dimensional net-effect matrix.}

\item{theta}{p-dimensional vector parameter.}
}
\value{
A \eqn{n \times n}{n by n} dimensional stochastic covariance matrix.
}
\description{
This function returns the stochastic covariance matrix W associated to a design matrix M, a net-effect matrix V, and a vector parameter theta.
}
\keyword{internal}
