% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{KLDiv}
\alias{KLDiv}
\title{Kullback-Leibler divergence}
\usage{
KLDiv(mu1, S1, mu2, S2)
}
\arguments{
\item{mu1}{mean vector of P.}

\item{S1}{covariance matrix of P.}

\item{mu2}{mean vector of Q.}

\item{S2}{covariance matrix of Q.}
}
\value{
Kullback-Leibler divergence \eqn{KL(P \Vert Q)}{KL(P || Q)} of Q from P.
}
\description{
Kullback-Leibler divergence \eqn{KL(P \Vert Q)}{KL(P || Q)} of Q from P for multivariate normal distributions
}
\keyword{internal}
