% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ldet}
\alias{ldet}
\title{Matrix determinant}
\usage{
ldet(A, logarithm = TRUE)
}
\arguments{
\item{A}{a matrix}

\item{logarithm}{logical; if TRUE (default) return the logarithm of the modulus of the determinant.}
}
\value{
the determinant of A.
}
\description{
Compute the determinant of a matrix.
}
\keyword{internal}
