% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{relErr}
\alias{relErr}
\title{Multivariate relative error}
\usage{
relErr(x, y)
}
\arguments{
\item{x}{a vector.}

\item{y}{a vector.}
}
\value{
the multivariate relative error between x and y.
}
\description{
Compute the multivariate relative error between two vectors.
}
