% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rmvNorm}
\alias{rmvNorm}
\title{Random generator from a Multivariate Normal distribution}
\usage{
rmvNorm(d, p, Mu, Sigma)
}
\arguments{
\item{d}{sample d.}

\item{p}{vector dimension.}

\item{Mu}{mean vector.}

\item{Sigma}{covariance matrix.}
}
\value{
A \eqn{p \times d}{p by d} matrix. Each column is a p-variate random vector from
a multivariate normal with mean vector Mu and covariance matrix Sigma.
}
\description{
This function generate a random vector from a multivariate normal distribution.
}
\keyword{internal}
