% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rndEffModelStats}
\alias{rndEffModelStats}
\title{Base and random-effects model statistics}
\usage{
rndEffModelStats(
  theta_null,
  theta_rndEff,
  V,
  M,
  M_bdiag,
  y,
  VCNs,
  nObs,
  trace = TRUE
)
}
\arguments{
\item{theta_null}{the estimated p-dimensional vector parameter for the base (null) model.}

\item{theta_rndEff}{the estimated p-dimensional vector parameter for the random-effects model.}

\item{M}{A \eqn{n \times K}{n by K} dimensional (design) matrix.}

\item{M_bdiag}{A\eqn{n \times Jp}{n by J*p} dimensional block-diagonal design matrix.
Each j-th block (\eqn{j = 1,\dots,J}{j = 1,..,J}) is a \eqn{n_j \times p}{n_j by p} dimensional design matrix for the j-th clone.}

\item{y}{n-dimensional vector of the time-adjacent cellular increments}

\item{VCNs}{A n-dimensional vector including values of the vector copy number corresponding to the cell counts of y.}

\item{nObs}{A K-dimensional vector including the frequencies of each clone k (\eqn{k = 1,\dots,K}{k = 1,..,K}).}

\item{trace}{Non-negative integer. If positive, tracing information on the progress of the optimization is produced.
This parameter is also passed to the optim() function.
Higher values may produce more tracing information: for method "L-BFGS-B" there are six levels of tracing.
(To understand exactly what these do see the source code: higher levels give more detail.)}
}
\value{
A vector of statistics associated to the fitted base and random-effects models:
\itemize{
\item{"nPar"}{number of parameters of the base(null) model}
\item{"cll"}{value of the conditional log-likelihood, in this case just the log-likelihood}
\item{"mll"}{value of the marginal log-likelihood, in this case just the log-likelihood}
\item{"cAIC"}{conditional Akaike Information Criterion (cAIC), in this case simply the AIC.}
\item{"mAIC"}{marginal Akaike Information Criterion (mAIC), in this case simply the AIC.}
\item{"Chi2"}{value of the \eqn{\chi^2}{Chi-squared} statistic \eqn{(y - M\theta)'S^{-1}(y - M\theta)}{(y - Mθ)'S^-1(y - Mθ)}.}
\item{"p-value"}{p-value of the \eqn{\chi^2}{Chi-squared} test for the null hypothesis that Chi2
follows a \eqn{\chi^2}{Chi-squared} distribution with n - nPar degrees of freedom.}
\item{"KLdiv"}{Kullback-Leibler divergence of the random-effects model from the null model.}
\item{"KLdiv/N"}{Rescaled Kullback-Leibler divergence of the random-effects model from the null model.}
\item{"BhattDist_nullCond"}{Bhattacharyya distance between the random-effects model and the null model.}
\item{"BhattDist_nullCond/N"}{Rescaled Bhattacharyya distance between the random-effects model and the null model.}
}
}
\description{
Main statistics from the fitted base and random-effects models
}
\keyword{internal}
