% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_RM}
\alias{Y_RM}
\title{Rhesus Macaque clonal tracking dataset}
\format{
A list containing clonal tracking data for each animal (ZH33, ZH17, ZG66).
Each clonal tracking dataset is a 3-dimensional array whose dimensions identify
\describe{
\item{1}{time, in months}
\item{2}{cell types: T, B, NK, Macrophages(M) and Granulocytes(G)}
\item{3}{unique barcodes (clones)}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3979461/bin/NIHMS567927-supplement-02.xlsx}
}
\usage{
Y_RM
}
\description{
A dataset containing clonal tracking cell counts from a Rhesus Macaque study.
}
\keyword{datasets}
